+++ app/assets/javascripts/components/core/Row/Row.js.jsx
@@ -1,13 +1,11 @@
 const React = require('react');
+const _     = require("underscore");
 
-const Row = React.createClass({
-  render() {
-    return (
-      <div className="row">
-        {this.props.children}
-      </div>
-    );
-  },
-});
+const Row = (props) => {
+  // We have to clone props because it's frozen (i.e. immutable):
+  const newProps     = _.clone(props);
+  newProps.className += " row";
+  return <div {...newProps} />;
+};
 
 module.exports = Row;