+++ db/migrate/20160923100253_add_partner_to_offers.rb
@@ -1,0 +1,6 @@
+class AddPartnerToOffers < ActiveRecord::Migration[5.0]
+  def change
+    add_column :offers, :partner, :integer
+    add_index  :offers, :partner
+  end
+end