+++ spec/features/travel_plans/index_page_spec.rb
@@ -19,9 +19,11 @@ describe "travel plans index page" do
     @eu  = create(:region, name: "Europe")
     @uk  = create(:country, name: "UK",     parent: @eu)
     @fr  = create(:country, name: "France", parent: @eu)
-    @lhr = create_airport("London Heathrow", :LHR, @uk)
-    @lgw = create_airport("London Gatwick",  :LGW, @uk)
-    @cdg = create_airport("Paris",           :CDG, @fr)
+    @lon = create(:city,    parent: @uk)
+    @par = create(:city,    parent: @fr)
+    @lhr = create_airport("London Heathrow", :LHR, @lon)
+    @lgw = create_airport("London Gatwick",  :LGW, @lon)
+    @cdg = create_airport("Paris",           :CDG, @par)
 
     @tp_single = account.travel_plans.create!(
       acceptable_classes:   [:economy, :premium_economy],